Task/integration

+++ test/unit/mongoDBUtils.js
@@ -29,8 +29,11 @@ function cleanDb(host, name, callback) {
     var url = 'mongodb://' + host + ':27017/' + name;
 
     MongoClient.connect(url, function(err, db) {
-        db.dropDatabase();
-        db.close();
+        if (db) {
+            db.dropDatabase();
+            db.close();
+        }
+
         callback();
     });
 }
